Re-work RaySpace layout
1. RaySpaceObject becomes ObjectRaySpace.
2. RaySpace is no longer contained in an ObjectRaySpace, so there is no
   unbounded tree of objects.
3. Partition the space containing obects with a KDTree, though it is not
   yet used.
4. Due to lagging, SDL joystick events no longer cause the locations of
   objects to be updated... this is only done at the start of each
   render. This removes a mutex... which I believe is okay, given that
   joystick events only update a shared 'input' structure, whose
   primitive values (floats) do not depend on one another.
5. A BitString problem was detected - 64 bit machines use size_t as 64
   bits... and bit shifts are only valid between 0 and 31... so the
   BitString type becomes an unsigned int.

bitstring.h  (5)
cli.c  (2,3)
compute.c  (1,2,3)
motion.c  (1,2)
testcase.c  (1,2)

gui.c
* main()  (4)
* sdl_main()  (4)
* render_RaySpace()  (4)
 + Mutex removed, no more joystick events moving racers.

raytrace.c/.h  (1,2,3)
* Del MultiCastRayParams
 + and associated functions. OpenCL code uses this but will not when it
   is brought back to working order.

space.c
* Replace memcpy()/memset() with loops. Loops are faster here... or at
  least make these set functions not appear in gprof output.

+++ bitstring.h
@@ -5,7 +5,7 @@
 
 #include "util.h"
 
-typedef size_t bitstring_t;
+typedef unsigned int bitstring_t;
 typedef bitstring_t BitString;
 
 BitString*
